rebase, address comments
