+++ system-tests/projects/qwik-app/cypress-ct-qwik/index.mjs
@@ -1,0 +1,3 @@
+export default function mount () {
+  return 'Legit mount function'
+}
