+++ system-tests/projects/qwik-app/cypress-ct-qwik/package.json
@@ -1,5 +1,9 @@
 {
   "name": "cypress-ct-qwik",
   "version": "1.0.0",
-  "main": "./index.js"
+  "main": "index.js",
+  "exports": {
+    "node": "./definition.cjs",
+    "default": "./index.mjs"
+  }
 }
